Improve compatibility for screen dimensions > 255.

+++ src/interpreter/fizmo.c
@@ -737,7 +737,7 @@ void ensure_dot_fizmo_dir_exists()
 #endif // DISABLE_CONFIGFILES
 
 
-void fizmo_new_screen_size(uint8_t width, uint8_t height)
+void fizmo_new_screen_size(uint16_t width, uint16_t height)
 {
   if (!z_mem)
   {
